[PERMISSION-23]
- create a new Mutation to add a single user
- create a new Mutation to update a single user
- optimize code / structure
