+++ node/resolvers/Mutations/Users.ts
@@ -3,7 +3,155 @@ import { currentSchema } from '../../utils'
 
 const config: any = currentSchema('b2b_users')
 
-export const saveUser = async (_: any, params: any, ctx: Context) => {
+const createPermission = async ({ lm, masterdata, vbase, params }: any) => {
+  const {
+    roleId,
+    canImpersonate,
+    name,
+    email,
+    userId,
+    clId,
+    orgId,
+    costId,
+    id,
+  } = params
+
+  let UserId = userId
+  let mdId = id
+
+  if (canImpersonate) {
+    await lm.saveUser(name, email).catch((err: any) => {
+      throw new Error(err)
+    })
+    const user = await lm.getUserIdByEmail(email)
+
+    UserId = user ?? userId
+  } else {
+    await lm.deleteUser(userId).catch((err: any) => {
+      throw new Error(err)
+    })
+  }
+
+  // check if new user's email already exists in storefront-permissions MD
+  if (!mdId) {
+    mdId = await masterdata
+      .searchDocuments({
+        dataEntity: config.name,
+        fields: ['id'],
+        pagination: { page: 1, pageSize: 90 },
+        schema: config.version,
+        where: `email=${email}`,
+      })
+      .then((r: any) => {
+        return r?.length ? r[0].id : null
+      })
+      .catch(() => {
+        return null
+      })
+  }
+
+  const ret = await masterdata
+    .createOrUpdateEntireDocument({
+      dataEntity: config.name,
+      fields: {
+        canImpersonate,
+        clId,
+        costId,
+        email,
+        name,
+        orgId,
+        roleId,
+        userId: UserId,
+      },
+      id: mdId,
+      schema: config.version,
+    })
+    .then((r: any) => {
+      return r
+    })
+    .catch((err: any) => {
+      if (err.response.status < 400) {
+        return {
+          DocumentId: id,
+        }
+      }
+
+      throw err
+    })
+
+  if (ret.DocumentId) {
+    await vbase.saveJSON('b2b_users', email, {
+      canImpersonate,
+      clId,
+      costId,
+      email,
+      id: ret.DocumentId,
+      name,
+      orgId,
+      roleId,
+      userId,
+    })
+  }
+}
+
+export const addUser = async (_: any, params: any, ctx: Context) => {
+  const {
+    clients: { masterdata, lm, vbase },
+    vtex: { logger },
+  } = ctx
+
+  try {
+    const newUser = await masterdata
+      .createDocument({
+        dataEntity: 'CL',
+        fields: {
+          email: params.email,
+          firstName: params.name,
+        },
+      })
+      .then((r: any) => {
+        return r
+      })
+      .catch((err: any) => {
+        if (err.response?.data?.Message === 'duplicated entry') {
+          return masterdata
+            .searchDocuments({
+              dataEntity: 'CL',
+              fields: ['id'],
+              pagination: {
+                page: 1,
+                pageSize: 1,
+              },
+              where: `email=${params.email}`,
+            })
+            .then((res: any) => {
+              return { DocumentId: res[0].id }
+            })
+        }
+
+        logger.error(err)
+        throw err
+      })
+
+    const cId = newUser.DocumentId
+
+    await createPermission({
+      lm,
+      masterdata,
+      params: {
+        ...params,
+        cId,
+      },
+      vbase,
+    })
+
+    return { status: 'success', message: '', id: cId }
+  } catch (err) {
+    return { status: 'error', message: err }
+  }
+}
+
+export const updateUser = async (_: any, params: any, ctx: Context) => {
   const {
     clients: { masterdata, lm, vbase },
     vtex: { logger },
@@ -12,132 +160,26 @@ export const saveUser = async (_: any, params: any, ctx: Context) => {
   try {
     // check if new user already exists in CL and create profile if not
     if (!params.clId) {
-      const newUser = await masterdata
-        .createDocument({
-          dataEntity: 'CL',
-          fields: {
-            firstName: params.name,
-            email: params.email,
-          },
-        })
-        .then((r: any) => {
-          return r
-        })
-        .catch((err: any) => {
-          if (err.response?.data?.Message === 'duplicated entry') {
-            return masterdata
-              .searchDocuments({
-                dataEntity: 'CL',
-                fields: ['id'],
-                where: `email=${params.email}`,
-                pagination: {
-                  page: 1,
-                  pageSize: 1,
-                },
-              })
-              .then((res: any) => {
-                return { DocumentId: res[0].id }
-              })
-          }
+      const { id: cId } = (await addUser(_, params, ctx)) as {
+        status: string
+        message: string
+        id: string
+      }
 
-          logger.error(err)
-          throw err
-        })
-
-      params.clId = newUser.DocumentId
+      params.cId = cId
     }
 
-    const {
-      roleId,
-      canImpersonate,
-      name,
-      email,
-      userId,
-      clId,
-      orgId,
-      costId,
-      id,
-    } = params
+    await createPermission({
+      lm,
+      masterdata,
+      params,
+      vbase,
+    })
 
-    let UserId = userId
-    let mdId = id
-
-    if (canImpersonate) {
-      await lm.saveUser(name, email).catch((err) => {
-        throw new Error(err)
-      })
-      const user = await lm.getUserIdByEmail(email)
-
-      UserId = user ?? userId
-    } else {
-      await lm.deleteUser(userId).catch((err) => {
-        throw new Error(err)
-      })
-    }
-
-    // check if new user's email already exists in storefront-permissions MD
-    if (!mdId) {
-      mdId = await masterdata
-        .searchDocuments({
-          dataEntity: config.name,
-          fields: ['id'],
-          schema: config.version,
-          pagination: { page: 1, pageSize: 90 },
-          where: `email=${email}`,
-        })
-        .then((r: any) => {
-          return r?.length ? r[0].id : null
-        })
-        .catch(() => {
-          return null
-        })
-    }
-
-    const ret = await masterdata
-      .createOrUpdateEntireDocument({
-        dataEntity: config.name,
-        fields: {
-          roleId,
-          userId: UserId,
-          clId,
-          orgId,
-          costId,
-          canImpersonate,
-          name,
-          email,
-        },
-        id: mdId,
-        schema: config.version,
-      })
-      .then((r: any) => {
-        return r
-      })
-      .catch((err) => {
-        if (err.response.status < 400) {
-          return {
-            DocumentId: id,
-          }
-        }
-
-        throw err
-      })
-
-    if (ret.DocumentId) {
-      await vbase.saveJSON('b2b_users', email, {
-        id: ret.DocumentId,
-        roleId,
-        userId,
-        clId,
-        orgId,
-        costId,
-        canImpersonate,
-        name,
-        email,
-      })
-    }
-
-    return { status: 'success', message: '', id: ret.DocumentId }
+    return { status: 'success', message: '', id: params.clId }
   } catch (e) {
+    logger.error({ message: 'Error saving user', error: e })
+
     return { status: 'error', message: e }
   }
 }
@@ -229,4 +271,18 @@ export const impersonateUser = async (_: any, params: any, ctx: Context) => {
   } catch (e) {
     return { status: 'error', message: e }
   }
+}
+
+/**
+ *
+ * Persist the user in the database
+ *
+ * @deprecated
+ *
+ * @param _
+ * @param params
+ * @param ctx
+ */
+export const saveUser = async (_: any, params: any, ctx: Context) => {
+  return updateUser(_, params, ctx)
 }
